Used correct type for constants and used nint() for comparing a real_r8 with an integer (CVS 1.59)
git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/trunk@2553 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: ccd1d67ef224ad8bc3930af67d24121956fb6a68

+++ models/cam/model_mod.f90
@@ -2872,15 +2872,15 @@ else
       ! Since the hybrid coord system is pure pressure at the top levels, I'll ignore these for now.
       call get_val(p_surf, x, lon_index, lat_index, -1, KIND_SURFACE_PRESSURE, vstatus)
       call plevs_cam (p_surf, num_levs, pfull)
-      highest_obs_level = 1
+      highest_obs_level = 1.0_r8
       threshold = highest_obs_pressure_mb*100.0_r8
       do while ((pfull(nint(highest_obs_level))) < threshold )
-         highest_obs_level = highest_obs_level +1
+         highest_obs_level = highest_obs_level + 1.0_r8
       end do
       deallocate(pfull)
    end if
 
-   if (level < highest_obs_level) then
+   if (level < nint(highest_obs_level)) then
       ! Exclude from assimilation the obs above a user specified level
       ! but still calculate the expected obs.
       istatus = 2
